Make sure both 'settings' and 'builders' are in the YAML.
That is, don't fall back to empty defaults for them.  This
is important in case users provide their own YAML configuration,
but have a typo in them (like 'biulders' instead of 'builders').
With the default of an empty dict, the builders would just be
skipped completely, and could lead to hard-to-debug issues.
At least with this PR it would be louder.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>