Fix intersphinx generation.
When you have more than 1 package in a repository, rosdoc2
will automatically generate intersphinx mappings between the
packages in the repository.  Unfortunately, there is a bug
where two commas get put into the entries, which means that
sphinx throws an "invalid syntax" warning when trying to parse
it.  Fix that here by removing the first one, and just rely
on the join() later on to add commas as appropriate.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>