Check `exhale` support before registering md lexer
The `exhale_args` mapping should only be updated with the appropriate
Pygments markdown lexer if both markdown support and exhale support are
requested. As such, an additional check for `enable_exhale` is added
into the markdown support configuration block.

An alternative would be to invoke `ensure_global` on `exhale_args`, but
that may create a needless copy of `exhale_args` incase `enable_exhale`
is false.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>