Add alternative installation instructions.
rosdoc2 can also be installed from pypi, which would be useful on a
non-Debian / -Ubuntu platform like Fedora or macOS.

Since the doc jobs on the build farm currently use rosdoc2 @ main I also
included instructions for installing it from source via pip.

Both methods recommend using virtualenvs and link to upstream
documentation for doing so.

I'm not opposed to adding `pipx` instructions or hints as well since
this project includes tests for it, but `pipx` is not a tool that's
currently recommended by the Open Robotics Infrastructure project and
the preferred installation method will be the ROS apt repositories.